+++ .env
@@ -1,1 +1,1 @@
-REACT_APP_BRUK_AUTENTISERING_I_DEV=true
+REACT_APP_BRUK_AUTENTISERING_I_DEV=false
