+++ src/utils/søknad.ts
@@ -61,25 +61,6 @@ export const fraStringTilTall = (tallAvTypenStreng: string) => {
   return parsed;
 };
 
-export const verdiTilTekstsvar = (
-  verdi: string | Date | boolean,
-  intl: IntlShape
-) => {
-  if (typeof verdi === 'string') {
-    return verdi;
-  } else if (typeof verdi === 'boolean') {
-    if (verdi === true) {
-      return hentTekst('svar.ja', intl);
-    } else {
-      return hentTekst('svar.nei', intl);
-    }
-  } else if (verdi instanceof Date) {
-    return formatDate(verdi);
-  } else {
-    return null;
-  }
-};
-
 export const settLabelOgVerdi = (objekt: any, variabelTilLabel: any) => {
   const nyttObjekt: any = {};
